renamed ErrorTypes.js to error-types.js
